\name{coef.PeakSegFPOP_df}
\alias{coef.PeakSegFPOP_df}
\title{coef PeakSegFPOP df}
\description{Create a list of data tables describing PeakSegFPOP model and
data.}
\usage{\method{coef}{PeakSegFPOP_df}(object, 
    ...)}
\arguments{
  \item{object}{
}
  \item{\dots}{
}
}

\value{list of data tables with named elements segments, loss, data,
changes, peaks.}

\author{Toby Dylan Hocking}





