\name{coef.PeakSegFPOP_dir}
\alias{coef.PeakSegFPOP_dir}
\title{coef PeakSegFPOP dir}
\description{Compute changes and peaks to display/plot.}
\usage{\method{coef}{PeakSegFPOP_dir}(object, 
    ...)}
\arguments{
  \item{object}{object }
  \item{\dots}{\dots }
}

\value{model list with additional named elements peaks and changes.}

\author{Toby Dylan Hocking}





