\name{fread.first}
\alias{fread.first}
\title{fread first}
\description{Read the first line of a text file.}
\usage{fread.first(file.name, col.name.vec)}
\arguments{
  \item{file.name}{Name of file to read.}
  \item{col.name.vec}{Character vector of column names.}
}

\value{Data table with one row.}

\author{Toby Dylan Hocking}





