\name{fread.last}
\alias{fread.last}
\title{fread last}
\description{Read the last line of a text file.}
\usage{fread.last(file.name, col.name.vec)}
\arguments{
  \item{file.name}{Name of file to read.}
  \item{col.name.vec}{Character vector of column names.}
}

\value{Data table with one row.}

\author{Toby Dylan Hocking}





