\name{plot.PeakSegFPOP_dir}
\alias{plot.PeakSegFPOP_dir}
\title{plot PeakSegFPOP dir}
\description{Plot a PeakSeg model with attached data.}
\usage{\method{plot}{PeakSegFPOP_dir}(x, 
    ...)}
\arguments{
  \item{x}{x }
  \item{\dots}{\dots }
}

\value{a ggplot.}

\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre]}





