\name{summary.PeakSegFPOP_dir}
\alias{summary.PeakSegFPOP_dir}
\title{summary PeakSegFPOP dir}
\description{Summary of \code{\link{PeakSegFPOP_dir}} \code{object}.}
\usage{\method{summary}{PeakSegFPOP_dir}(object, 
    ...)}
\arguments{
  \item{object}{object }
  \item{\dots}{\dots }
}

\value{Data table with one row and columns describing model summary.}

\author{Toby Dylan Hocking}





