\name{plot.PeakSegFPOP_df}
\alias{plot.PeakSegFPOP_df}
\title{plot PeakSegFPOP df}
\description{Plot a PeakSeg model with attached data.}
\usage{\method{plot}{PeakSegFPOP_df}(x, 
    ...)}
\arguments{
  \item{x}{x }
  \item{\dots}{\dots }
}

\value{a ggplot.}

\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre]}





